\name{swan-package}
\alias{swan-package}
\alias{swan}
\docType{package}
\title{
	Statistical Structrual Variant Analyzer (SWAN)
}
\description{
  SWAN is a Statistical Structrual Variant Analyzer (SWAN) for structrual variant
  calling based on next generation sequencing data and a joint log likelihood
  scan statistic using mixed heterogenous marked Poisson process models.
}
\details{
  \tabular{ll}{
    Package: \tab swan\cr
    Type: \tab Package\cr
    Version: \tab 1.0.0\cr
    Date: \tab 2015-06-01\cr
    License: \tab Creative Commons BY-NC-SA\cr
  }

	see \url{http://bitbucket.org/charade/swan}\cr
  see packageDescription('swan') for more details\cr

}
\author{
  Author: Charlie Xia <li.xia@stanford.edu>
}
\references{
	Scanning a Poisson random field for local signals. 
	Nancy R Zhang, Benjamin Yakir, Charlie L Xia, David Siegmund.
	arXiv preprint arXiv:1406.3258
	\url{http://arxiv.org/abs/1406.3258}
}
\note{
	\link{get_MPRs}
}
\keyword{ package }
\examples{
  packageDescription('swan')
}
