% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swan.R
\name{swan}
\alias{swan}
\title{Run SWaN}
\usage{
swan(df, sample_rate)
}
\arguments{
\item{df}{A \code{data.frame} with \code{X/Y/Z} and a time in
\code{HEADER_TIMESTAMP}/\code{HEADER_TIME_STAMP}.}

\item{sample_rate}{the sampling rate of the data}
}
\value{
A \code{data.frame} of predictions.
}
\description{
Run SWaN
}
\examples{
if (swan_check()) {
  file = system.file("extdata", "73557_subset.csv.gz", package = "swan")
  if (requireNamespace("readr", quietly = TRUE)) {
    df = readr::read_csv(file)
    readr::stop_for_problems(df)
  } else {
    df = read.csv(file)
    df = df \%>\%
      dplyr::mutate(
        HEADER_TIMESTAMP = sub("Z$", "", HEADER_TIMESTAMP),
        HEADER_TIMESTAMP = strptime(HEADER_TIMESTAMP,
                                    "\%Y-\%m-\%dT\%H:\%M:\%OS")) \%>\%
      dplyr::as_tibble()
  }
  out = swan(df, sample_rate = 80L)
  out$second_pass \%>\% dplyr::count(prediction)
}
}
