% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_create_swan.R
\name{conda_create_swan}
\alias{conda_create_swan}
\alias{have_swan_condaenv}
\title{Create Conda Environmment for \code{swan}}
\usage{
conda_create_swan(envname = "swan", ...)

have_swan_condaenv(envname = "swan")
}
\arguments{
\item{envname}{conda environment name for the modules to be installed}

\item{...}{additional arguments to pass to \code{\link[reticulate:conda-tools]{reticulate::conda_create()}}}
}
\value{
Nothing, called for side effects
}
\description{
Create Conda Environmment for \code{swan}
}
